+++ Core/Charging/ChargingModel/Chargepoint/IChargingPoint.cs
@@ -1,8 +1,5 @@
 namespace Core.Charging.ChargingModel.Chargepoint;
 
-using System.Collections.Immutable;
-using Core.Shared;
-
 /// <summary>
 /// Marker interface for all charging point types.
 /// Use <see cref="ISingleChargingPoint"/> or <see cref="IDualChargingPoint"/>.
