+++ Engine/Metrics/Snapshots/StationSnapshotMetrics.cs
@@ -1,8 +1,5 @@
 namespace Engine.Metrics.Snapshots;
 
-using Core.Charging;
-using Core.Shared;
-
 /// <summary>
 /// A point-in-time snapshot of a single station's metrics.
 /// Collected once per simulation hour.
